checkout correct branch in case of GITFLAGS are present
passing GITFLAGS is desired to speed up clone when making packages. possible
flags are --filter=tree:0 or --depth=1. makepkg makes a local clone subsequent,
which in some cases fail, or, is on HEAD instead of the desired commit. so make
sure the desired branch/tag/commit specified in PKGBUILD is checked out in case
of shallow or partial clone.

when doing the default mirroring, this is not necessary, as all
branches are on local disk.

in case the repository contains submodules, make sure they get cloned into
src directory.