Fix sync-update-remove problem by double checking
This is fix to tests/sync-update-package-removing-required-provides.py,
and it fixes tests/sync406.py as well.

Basically the problem was that if sync-updating removes required
provides, pacman would see these provides as miss depends and fail at
once. What I do is to add the depends back, prepare them and check
again.

Signed-off-by: Dale young <daleyoung4242@gmail.com>